feat(next): done - add metadata to convo pages
